webserver/requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.3
fastui==0.7.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
redis==5.1.1
setuptools==75.3.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
supervisor==4.2.5
typing_extensions==4.12.2
uvicorn==0.32.0
